reshuffle a bit
